feat: Transaction / Create TransactionOption component #176
